finalfinal

+++ main.sh
@@ -32,60 +32,140 @@ doAction () {	#Main function used to control the program
 		--help) #Prints all input commands
 			printMenu;;	
 		make)	#Creates a new repository as well as its relevant log file
-			createRepository "$3" $2	
-			createLogFile $(findRepoIndex $2);;
+			if [ $# -eq 3 ]; then
+				createRepository "$3" "$2"
+				createLogFile $(findRepoIndex "$2")
+			else 
+				echo "3 parameters needed."
+			fi;;
 		delete)	#Deletes an existing repository
-			deleteRepository "$3" $(findRepoIndex $2);;
+			if [ $# -eq 2 ]; then
+				deleteRepository $(findRepoIndex "$2")
+			else 
+				echo "2 parameters needed."
+			fi;;
 		repos)	#Prints available repositories to the console
-			printRepos;;
+			if [ $# -eq 1 ]; then
+				printRepos
+			else 
+				echo "1 parameter needed."
+			fi;;
 		createfile)	#Creates a new file
-			createNewFile "$3" $(findRepoIndex $2);;
+			if [ $# -eq 3 ]; then
+				createNewFile "$3" $(findRepoIndex "$2")
+			else 
+				echo "3 parameters needed."
+			fi;;
 		deletefile)	#Deletes an existing file
-			deleteFile "$3" $(findRepoIndex $2);;
+			if [ $# -eq 3 ]; then
+				deleteFile "$3" $(findRepoIndex "$2")
+			else 
+				echo "3 parameters needed."
+			fi;;
 		list)		#Lists all files in the current repository
-			listFiles $(findRepoIndex $2);;
+			if [ $# -eq 2 ]; then
+				listFiles $(findRepoIndex "$2")
+			else 
+				echo "2 parameters needed."
+			fi;;
 		edit)		#Opens a file externally for editing
-			editFile $3 $(findRepoIndex $2);;
+			if [ $# -eq 3 ]; then
+				editFile "$3" $(findRepoIndex "$2")
+			else 
+				echo "3 parameters needed."
+			fi;;
 		stage)		#Moves a file (or all files if -a is used) to the staging area of the repository
-			if [ "$3" = "-a" ]; then
-				moveAllToStagingFolder $(findRepoIndex $2)
-			else
-				moveToStagingFolder $3 $(findRepoIndex $2)
+			if [ $# -eq 3 ]; then
+				if [ "$3" = "-a" ]; then
+					moveAllToStagingFolder $(findRepoIndex "$2")
+				else
+					moveToStagingFolder $3 $(findRepoIndex "$2")
+				fi
+			else 
+				echo "3 parameters needed."
 			fi;;
 		unstage)	#Removes a file (or all files if -a is used) from the staging area of the repository
-			if [ "$3" = "-a" ]; then
-				clearStagingFolder $(findRepoIndex $2)
-			else
-				moveFromStagingFolder $3 $(findRepoIndex $2)
+			if [ $# -eq 3 ]; then
+				if [ "$3" = "-a" ]; then
+					clearStagingFolder $(findRepoIndex "$2")
+				else
+					moveFromStagingFolder $3 $(findRepoIndex "$2")
+				fi
+			else 
+				echo "3 parameters needed."
 			fi;;
 		stageclear)	#Does the same as does the same as unstage -a
-			clearStagingFolder $(findRepoIndex $2);;
+			if [ $# -eq 2 ]; then
+				clearStagingFolder $(findRepoIndex "$2")
+			else 
+				echo "2 parameters needed."
+			fi;;
 		stageall)	#Does the same as stage -a
-			moveAllToStagingFolder $(findRepoIndex $2);;
+			if [ $# -eq 2 ]; then
+				moveAllToStagingFolder $(findRepoIndex "$2")
+			else 
+				echo "2 parameters needed."
+			fi;;
 		commit)		#Commits a file from the staging area of the current repository
-			makeCommit $3 $(findRepoIndex $2);;
+			if [ $# -eq 3 ] || [ $# -eq 2 ] ; then
+				makeCommit "$3" $(findRepoIndex "$2")
+			else 
+				echo "2 or 3 parameters needed."
+			fi;;
 		revert)		#Reverts a specific commit
-			revertCommit $3 $(findRepoIndex $2);;
+			if [ $# -eq 3 ]; then
+				revertCommit "$3" $(findRepoIndex "$2")
+			else 
+				echo "3 parameters needed."
+			fi;;
 		commits)	#Prints all commits made in the current repository
-			printCommits $(findRepoIndex $2);;
+			if [ $# -eq 2 ]; then
+				printCommits $(findRepoIndex "$2")
+			else 
+				echo "2 parameters needed."
+			fi;;
 		zip)
-			zipRep $(findRepoIndex $2);;
-		setPassword)
-			setPassword $3 $2;;
+			if [ $# -eq 2 ]; then
+				zipRep $(findRepoIndex "$2")
+			else 
+				echo "2 parameters needed."
+			fi;;
+		setpassword)
+			if [ $# -eq 3 ]; then
+				setPassword "$3" "$2"
+			else 
+				echo "3 parameters needed."
+			fi;;
 		autobackup)
-			automaticBackups $(findRepoIndex $2) &
-			echo -e "Automatically backing up all repository files. To stop enter \"kill $!\"";;
+			if [ $# -eq 2 ]; then
+				automaticBackups $(findRepoIndex "$2") &
+				echo -e "Automatically backing up all repository files. To stop enter \"kill $!\""
+			else 
+				echo "2 parameters needed."
+			fi;;
 		autostaging)
-			automaticStaging $(findRepoIndex $2) &
-			echo -e "Automatically staging changed repository files. To stop enter \"kill $!\"";;
+			if [ $# -eq 2 ]; then
+				automaticStaging $(findRepoIndex "$2") &
+				echo -e "Automatically staging changed repository files. To stop enter \"kill $!\""
+			else 
+				echo "2 parameters needed."
+			fi;;
 		permission)
-			createUserGroup $(findRepoIndex $2)
-			lockToUserGroup $(findRepoIndex $2);;
+			if [ $# -eq 2 ]; then
+				createUserGroup $(findRepoIndex "$2")
+				lockToUserGroup $(findRepoIndex "$2")
+			else 
+				echo "2 parameters needed."
+			fi;;
 		allowuser)
-			createUserGroup $(findRepoIndex $2)
-			addUsersToGroup $3 $(findRepoIndex $2);;
+			if [ $# -eq 3 ]; then
+				createUserGroup $(findRepoIndex "$2")
+				addUsersToGroup $3 $(findRepoIndex "$2")
+			else 
+				echo "3 parameters needed."
+			fi;;
 		*)
-			echo "Error, unknown command";;
+			echo "Error, unknown command"
 	esac
 }
 
@@ -108,25 +188,26 @@ printMenu () {
 	echo -e "revert\t\trevert a commit"
 	echo -e "commits\t\tprints a list of existing commits"
 	echo -e "zip\t\tzip a repository"
+	echo -e "setpassword\tsetting a password for a repository"
 	echo -e "autobackup\tautomatically backing up files"
 	echo -e "autostaging\tautomatically staging edited files"
 	echo -e "permission\tpermission protection"
 	echo -e "allowuser\tassigning users and groups"
-	echo -e "exit\t\texits Jet\n"
+	echo
 }
 
 createRepository () {
 	#assumptions: $1 is a path to the repository, $2 is a repository name
 	cd $HOME
-	if [ -d ${1} ]; then
-		cd $1
-		mkdir .${2}
+	if [ -d "${1}" ]; then
+		cd "$1"
+		mkdir ".${2}"
 	else
 		echo "The path to the new Repository is invalid. Creating new pathway..."
-		mkdir -p $1/.$2
-		cd $1
+		mkdir -p "$1/.$2"
+		cd "$1"
 	fi
-	mkdir .${2}/${stagingFolder}
+	mkdir ".${2}/${stagingFolder}"
 	repositories[${#repositories[@]}]="$2"
 	repositoryPaths[${#repositoryPaths[@]}]="$1"
 }
@@ -139,49 +220,69 @@ createLogFile () {
 }
 
 deleteRepository () {
-	#assumptions: $1 is a repository name, $2 is a repository index
-	cd $HOME
-	if [ -d "${repositoryPaths[$2]}" ]; then
-		rm -r "${repositoryPaths[$2]}"
-		repositories[$2]=""
-		repositoryPaths[$2]=""
+	#assumptions: $1 is a repository index
+	if [ "$1" != "-1" ]; then
+		if [ "$(validatePassword "$(findPasswordIndex ${repositories["$2"]})")" = 0 ]; then
+			cd $HOME
+			if [ -d "${repositoryPaths[$1]}" ]; then
+				rm -r "${repositoryPaths[$1]}"
+				repositories[$1]=""
+				repositoryPaths[$1]=""
+			else
+				echo "The repository you're trying to delete doesn't exist"
+			fi
+		fi
 	else
-		echo "The repository you're trying to delete doesn't exist"
+		echo "No repository with this name found"
 	fi
 }
 
 printRepos () {
-	for i in ${!repositories[@]}; do
-		echo -e "${repositories[$i]}\t${repositoryPaths[$i]}"
-  	done
+	if [ ${#repositories[@]} -gt 0 ]; then
+		for i in ${!repositories[@]}; do
+			echo -e "${repositories[$i]}\t${repositoryPaths[$i]}"
+	  	done
+	else
+		echo "No repositories found"
+	fi 
 }
 
 createNewFile () {
 	#assumptions: $1 is a filename, $2 repository index
-	cd $HOME
-	cd "./${repositoryPaths[$2]}"
-	if [[ "$1" == *"/"* ]]; then
-		mkdir -p "${1%/*}" && touch "$1"
-	else 
-		touch "$1"
+	if [ "$(validatePassword "$(findPasswordIndex ${repositories["$2"]})")" = 0 ]; then
+		cd $HOME
+		cd "./${repositoryPaths[$2]}"
+		if [[ "$1" == *"/"* ]]; then
+			mkdir -p "${1%/*}" && touch "$1"
+		else 
+			touch "$1"
+		fi
+		moveToStagingFolder "$1" $2
 	fi
-	moveToStagingFolder $1 $2
 }
 
 deleteFile () {
 	#assumptions: $1 is a filename, $2 repository index
-	cd $HOME
-	cd "./${repositoryPaths[$2]}"
-	rm -r "$1"
-	moveFromStagingFolder $1 $2
+	if [ "$2" != "-1" ]; then
+		if [ "$(validatePassword "$(findPasswordIndex ${repositories["$2"]})")" = 0 ]; then
+			cd $HOME
+			cd "./${repositoryPaths[$2]}"
+			rm -r "$1"
+			moveFromStagingFolder "$1" $2
+		fi
+	fi
 }
 
 listFiles () {
   	#assumptions: $1 is a repository index
 	cd $HOME
-  	if [ -d "${repositoryPaths[$1]}" ]; then
+  	if [ $1 -ge 0 ] && [ -d "${repositoryPaths[$1]}" ]; then
 		cd "./${repositoryPaths[$1]}"
-		ls
+  		if [ $(ls -1q | wc -l) -gt 0 ]; then
+			ls
+		else
+			echo "No files found"
+		fi
 	else
 		echo "The repository you're trying to list files from doesn't exist"
 	fi
@@ -192,9 +293,9 @@ editFile () {
 	cd $HOME
 	if [ -d "${repositoryPaths[$2]}" ]; then
 		cd "./${repositoryPaths[$2]}"
-		if [ -f $1 ]; then
-			if [ "$(validatePassword "$(findPasswordIndex ${repositories[$2]})")" = 0 ]; then
-				xdg-open $1
+		if [ -f "$1" ]; then
+			if [ "$(validatePassword "$(findPasswordIndex ${repositories["$2"]})")" = 0 ]; then
+				xdg-open "$1"
 			fi
 		else
 			echo "The file you're trying to open doesn't exist"
@@ -210,7 +311,7 @@ moveToStagingFolder () {
 	if [ -d "${repositoryPaths[$2]}" ]; then
 		cd "./${repositoryPaths[$2]}"
 		if [ -f "$1" ]; then
-			cp -r ${1} ./.${repositories[$2]}/${stagingFolder}/
+			cp -r "${1}" "./.${repositories[$2]}/${stagingFolder}/"
 		else
 			echo "The file you're trying to stage doesn't exist"
 		fi
@@ -225,10 +326,10 @@ moveToStagingFolder () {
 moveFromStagingFolder () {
 	#$1 filename, $2 rep index
 	cd $HOME
-	if [ -d ${repositoryPaths[$2]} ]; then
-		cd ./${repositoryPaths[$2]}/.${repositories[$2]}/${stagingFolder}
-		if [ -f $1 ]; then
-			rm -r $1
+	if [ -d "${repositoryPaths[$2]}" ]; then
+		cd "./${repositoryPaths[$2]}/.${repositories[$2]}/${stagingFolder}"
+		if [ -f "$1" ]; then
+			rm -r "$1"
 		else
 			echo "The file you're trying to remove from the staging area doesn't exist"
 		fi
@@ -270,18 +371,19 @@ makeCommit () {
 	#assumptions: $2 is a repository index, $1 is a commit message
 	#--------------if [ repo doesnt exist ] then print error, if [ no commit message given ] then print error, else print code below
 	cd $HOME
-	if [ -d ${repositoryPaths[$2]} ] && [ -n $1 ]; then
+	if [ -d "${repositoryPaths[$2]}" ] && [ -n "$1" ]; then
 	  local date=$(date +'%Y-%m-%d %H:%M:%S')
 		local timestamp=$(date +%s)
 		cd $HOME
-		cd ./${repositoryPaths[$2]}/.${repositories[$2]}
+		cd "./${repositoryPaths[$2]}/.${repositories[$2]}"
 		if [ $(ls -1q ./${stagingFolder} | wc -l) -gt 0 ]; then
-	  	addCommitToLogFile $2 $1 "$date" $timestamp
+	  	addCommitToLogFile "$2" "$1" "$date" $timestamp
 			mkdir $timestamp
 			mv ./${stagingFolder}/* ./$timestamp
 		else
 			echo "No files have been staged yet."
 		fi
+		moveAllToStagingFolder "$2"
 	elif [ -d $2 ]; then
 		echo "The repository you're trying to commit doesn't exist."
 	else
@@ -292,22 +394,25 @@ makeCommit () {
 revertCommit () {
 	#assumptions: $2 is a repository index, $1 is a commit timestamp
 	cd $HOME
-	if [ -d ${repositoryPaths[$2]} ]; then
-	  local date=$(date +'%Y-%m-%d %H:%M:%S')
-		addCommitToLogFile $2 "Reverted commit $1" $timestamp
-		cd $HOME
-		cd ./${repositoryPaths[$2]}
-		for i in *; do
-			if [ "$i" != .$repositories[$2] ] && [ "$i" != "*" ]; then
-				rm -r "$i"
-			fi
-		done
-		cd ./.${repositories[$2]}/$1
-		for i in *; do
-			if [ "$i" != "*" ]; then
-				cp -r "$i" ../..
-			fi
-		done
+	if [ -d "./${repositoryPaths[$2]}/.${repositories[$2]}/$1" ]; then
+		if [ "$(validatePassword "$(findPasswordIndex ${repositories["$2"]})")" = 0 ]; then
+		  	local date=$(date +'%Y-%m-%d %H:%M:%S')
+			addCommitToLogFile "$2" "Reverted commit $1" "$date" $timestamp
+			cd $HOME
+			cd "./${repositoryPaths[$2]}"
+			for i in *; do
+				if [ "$i" != ".$repositories[$2]" ] && [ "$i" != "*" ]; then
+					rm -r "$i"
+				fi
+			done
+			cd "./.${repositories[$2]}/$1"
+			for i in *; do
+				if [ "$i" != "*" ]; then
+					cp -r "$i" ../..
+				fi
+			done
+			moveAllToStagingFolder "$2"
+		fi
 	else
 		echo "The commit you're trying to revert hasn't been made"
 	fi
@@ -316,8 +421,8 @@ revertCommit () {
 printCommits () {
 	#$1 rep index
 	cd $HOME
-	if [ -d $1 ]; then
-		cd ./${repositoryPaths[$1]}/.${repositories[$1]}/
+	if [ -d "./${repositoryPaths[$1]}/.${repositories[$1]}/" ]; then
+		cd "./${repositoryPaths[$1]}/.${repositories[$1]}/"
 		for i in *; do
 			if [ "$i" != $stagingFolder ] && [ "$i" != $logFile ] && [ "$i" != "*" ]; then
 				echo "$i"
@@ -331,8 +436,8 @@ printCommits () {
 zipRep () {
 	#assumptions: $1 is a repository index
 	cd $HOME
-	if [ -d ${repositoryPaths[$1]} ]; then
-		zip -r ./${repositoryPaths[$1]}/${repositories[$1]}.zip ./${repositoryPaths[$1]}
+	if [ -d "${repositoryPaths[$1]}" ]; then
+		zip -r "./${repositoryPaths[$1]}/${repositories[$1]}.zip" "./${repositoryPaths[$1]}"
 	else
 		echo "The repository you're trying to compress doesn't exist"
 	fi
@@ -340,7 +445,7 @@ zipRep () {
 
 setPassword () {
 	#assumptions: $1 is a password, $2 is a repository name
-	passwordRepos[${#passwordRepos[@]}]=$2
+	passwordRepos[${#passwordRepos[@]}]="$2"
 	passwords[${#passwords[@]}]="$1"
 }
 
@@ -417,7 +522,7 @@ findRepoIndex () {
 addCommitToLogFile () {
 	#assumptions: $1 is a repository index, $2 is a commit message, $3 is date, $4 is commit timestamp (optional), log file exists
 	cd $HOME
-	cd ./${repositoryPaths[$1]}/.${repositories[$1]}
+	cd "./${repositoryPaths[$1]}/.${repositories[$1]}"
 	echo -e "${3}\t${2}\t$(whoami)\t${4}" >> ${logFile}
 }
 
@@ -425,7 +530,7 @@ addCommitToLogFile () {
 
 saveConfig () {
 	cd $HOME
-	rm $configFile
+	rm $configFile 2> /dev/null
 	touch $configFile 
 	for i in ${!repositories[@]}; do
 		if [ "${repositories[$i]}" != "" ]; then
@@ -465,7 +570,7 @@ loadConfig () {
 	fi
 	if [ -f "$passwordsFileEncrypted" ]; then
 		# echo "Enter an ecnryption password"
-		gpg --passphrase "test" --batch -d $passwordsFileEncrypted > $passwordsFile
+		gpg --passphrase "test" --batch -d $passwordsFileEncrypted > $passwordsFile 2> /dev/null
 		odd=true
 		while read line; do
 			if [ $odd = true ]; then
@@ -512,23 +617,23 @@ validatePassword () {
 }
 
 #################
+#$1 - command name
+#$2 - repository name (not case sensitive)
+#$3... - function arguments
+
 
 # Validation for general 
 if [ $# -eq 0 ]; then
 	#printMenu	-	leave commented until finished for ease of testing purposes
 	echo "No arguments were given."
-fi
-
-#$1 - command name
-#$2 - repository name (not case sensitive)
-#$3... - function arguments
-
-loadConfig
-doAction "$@"
-if [ "$1" != "autoBackup" ]; then
-	saveConfig
-fi
-if ! [ -z $2 ] && [ $1 != "list" ]; then
-	echo Files:
-	listFiles $(findRepoIndex $2)
+else
+	loadConfig
+	doAction "$@"
+	if [ "$1" != "autoBackup" ]; then
+		saveConfig
+	fi
+	if ! [ -z "$2" ] && [ "$1" != "list" ] && [ "$1" != "--help" ]; then
+		echo Files:
+		listFiles $(findRepoIndex "$2")
+	fi
 fi
